feat: improve project structure
